Replace the Screenshot item with a direct access to Sound settings
The screenshot button is never used as you can get the same functionality
with the PrintScr key, so replace it with a direct access to the sound
panel of the control center, which I use much more often (e.g. whenever
changing the sound input/output from speakers to a headset for calls).